fix and clean validation
