+++ presentation/src/main/java/com/mibaldi/presentation/ui/common/ValidateUtils.kt
@@ -1,0 +1,45 @@
+package com.mibaldi.presentation.ui.common
+
+import android.text.TextUtils
+import android.util.Patterns
+
+sealed class Field {
+    class Email(val error:String?) : Field()
+    class Password(val error:String?): Field()
+}
+
+
+fun validatePassword(password: String, valid: Boolean,myFunction:(Field)-> Unit): Boolean {
+    var valid1 = valid
+    when {
+        TextUtils.isEmpty(password) -> {
+            myFunction(Field.Password("Required."))
+            valid1 = false
+        }
+        password.length < 6 -> {
+            myFunction(Field.Password("The password should be at least 6 character."))
+            valid1 = false
+        }
+        else -> {
+            myFunction(Field.Password(null))
+        }
+    }
+    return valid1
+}
+
+fun validateEmail(email: String, valid: Boolean,myFunction:(Field)-> Unit): Boolean {
+    var valid1 = valid
+    when {
+        TextUtils.isEmpty(email) -> {
+            myFunction(Field.Email("Required"))
+
+            valid1 = false
+        }
+        !Patterns.EMAIL_ADDRESS.matcher(email).matches() -> {
+            myFunction(Field.Email("The email address is badly formatted."))
+            valid1 = false
+        }
+        else -> myFunction(Field.Email(null))
+    }
+    return valid1
+}
